+++ tools/PI/DevHome.PI/PIApp.xaml.cs
@@ -68,7 +68,6 @@ public partial class App : Application, IApp
                 services.AddSingleton<ResourceUsagePage>();
                 services.AddSingleton<ResourceUsagePageViewModel>();
                 services.AddSingleton<WERPage>();
-                services.AddSingleton<WERPageViewModel>();
                 services.AddSingleton<WinLogsPage>();
                 services.AddSingleton<WinLogsPageViewModel>();
                 services.AddSingleton<SettingsPage>();
